+++ lib/syscalls/open.c
@@ -1,7 +1,7 @@
 #include "syscall_preamble.h"
 #include <sys/types.h>
 long open(const char *path, int flags, mode_t mode) {
-  int ret;
+  long ret;
   syscall_output()
   syscall_input(open, "D"(path), "S"(flags), "d"(mode))
   syscall_clobber()
